+++ packages/create-sitecore-jss/src/templates/angular/src/app/JssState.ts
@@ -1,12 +1,8 @@
-import { RouteData, LayoutServiceContextData } from '@sitecore-jss/sitecore-jss-angular';
+import { BaseJssState } from '@sitecore-jss/sitecore-jss-angular';
 import { LayoutServiceError } from './layout/jss-layout.service';
 
-export class JssState {
+export class JssState extends BaseJssState {
   language: string;
   serverRoute: string;
   routeFetchError?: LayoutServiceError;
-  sitecore?: LayoutServiceContextData & {
-    route: RouteData | null;
-  };
-  viewBag: { [key: string]: unknown };
 }