+++ packages/sitecore-jss/src/editing/index.ts
@@ -1,5 +1,6 @@
 export { GraphQLEditingService } from './graphql-editing-service';
 export {
+  DEFAULT_PLACEHOLDER_UID,
   ExperienceEditor,
   HorizonEditor,
   isEditorActive,