+++ packages/sitecore-jss/src/editing/utils.ts
@@ -1,6 +1,11 @@
 import isServer from '../utils/is-server';
 
 /**
+ * Default value of uid for root placeholder when uid is not present.
+ */
+export const DEFAULT_PLACEHOLDER_UID = '00000000-0000-0000-0000-000000000000';
+
+/**
  * Query parameter for editing secret
  */
 export const QUERY_PARAM_EDITING_SECRET = 'secret';